+++ StratEngine/UCIHandler.cpp
@@ -180,6 +180,7 @@ void UciHandler::cmd_eval()
         send(eval_term_row("mopup",    terms.mopup[WHITE],    terms.mopup[BLACK]));
         send(eval_term_row("bishops",  terms.bishops[WHITE],  terms.bishops[BLACK]));
         send(eval_term_row("castling", terms.castling[WHITE], terms.castling[BLACK]));
+        send(eval_term_row("mobility", terms.mobility[WHITE], terms.mobility[BLACK]));
         send(rule);
 
         // The sum of the net column. It must equal the 'white pov' line below;
@@ -191,7 +192,8 @@ void UciHandler::cmd_eval()
                           + (terms.pst[WHITE]      - terms.pst[BLACK])
                           + (terms.mopup[WHITE]    - terms.mopup[BLACK])
                           + (terms.bishops[WHITE]  - terms.bishops[BLACK])
-                          + (terms.castling[WHITE] - terms.castling[BLACK]);
+                          + (terms.castling[WHITE] - terms.castling[BLACK])
+                          + (terms.mobility[WHITE] - terms.mobility[BLACK]);
         const std::string sum_label = "sum (white pov)";
         send(sum_label + pad_left(std::to_string(net_sum),
                                   EVAL_TABLE_WIDTH - static_cast<int>(sum_label.size())));